refactor: Update input type in AddProduct component to fix rating input issue
