+++ Ecomerce/src/components/AddProduct.tsx
@@ -86,7 +86,7 @@ export default function AddProduct() {
                         Rating
                     </label>
                     <input
-                        type='number'
+                        type='text'
                         {...register('rating')}
                         id='rating'
                         className='bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-600 dark:border-gray-500 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500'
