+++ Ecomerce/src/hooks/Products/usePostProducts.ts
@@ -1,14 +1,15 @@
-import { ChangeEvent, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { createProduct, getAllProducts } from '../../services/Products'
 import { productType } from '../../types/Product'
 import { useNavigate } from 'react-router-dom'
+import useImages from '../useImages'
 
 export default function usePostProducts() {
 
     const { register, handleSubmit, setValue } = useForm()
+    const { images, handleImageChange, addImageField } = useImages()
     const navigate = useNavigate()
-    const [images, setImages] = useState([''])
     const [nextProductId, setNextProductId] = useState(0)
     
     useEffect(() => {
@@ -38,14 +39,5 @@ export default function usePostProducts() {
         }
     })
 
-    const handleImageChange = (index: number, event: ChangeEvent<HTMLInputElement>) => {
-        const newImages = [...images]
-        newImages[index] = event.target.value
-        setImages(newImages)
-    }
-
-    const addImageField = () => {
-        setImages([...images, ''])
-    }
     return { register, OnSubmit, nextProductId, handleImageChange, addImageField, images}
 }
