+++ Ecomerce/src/hooks/Products/usePutProduct.ts
@@ -1,15 +1,16 @@
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { productType } from "../../types/Product";
-import { ChangeEvent, useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useGetProdutById } from "./useGetProductById";
 import { updateProduct } from "../../services/Products";
+import useImages from "../useImages";
 
 export default function usePutProduct(categoryId: string, productId: string) {
   const { register, handleSubmit, setValue } = useForm();
   const navigate = useNavigate();
   const { product } = useGetProdutById(categoryId, productId);
-  const [images, setImages] = useState<string[]>([""]);
+  const { images, setImages, handleImageChange, addImageField } = useImages();
 
   useEffect(() => {
     setValue("id", product?.id);
@@ -37,19 +38,6 @@ export default function usePutProduct(categoryId: string, productId: string) {
       console.error("Error uptating product:", error);
     }
   });
-
-  const handleImageChange = (
-    index: number,
-    event: ChangeEvent<HTMLInputElement>
-  ) => {
-    const newImages = [...images];
-    newImages[index] = event.target.value;
-    setImages(newImages);
-  };
-
-  const addImageField = () => {
-    setImages([...images, ""]);
-  };
 
   return { register, OnSubmit, images, handleImageChange, addImageField };
 }
